pyproject.toml
[tool.poetry]
name = "sql-custom-types"
version = "0.1.0"
description = ""
authors = ["Marco Visani <92217074+mvisani@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
